Add username check when loading a session from cookies
This change sets a cookie containing user_id on LTI connect. If the
session dies, pages should still be able to be viewed by using the
session token and username stored in the cookie.

If the user logs in as a different user and connects to the LTI tool,
the username cookie will be overridden, invalidating the original user's
cookies which are now username checked too.

This stops two users on the same machine sharing notes via login/logout.