Detect Safari ITP and show a message requesting access
This will need reworking in future to detect earlier in the loading
of the page and implementing as it's own LTIPage class.

However, as a temporary patch this seems to get the job done for the
moment so that users on MacOS Catalina & Safari can access content in
an iframe.