Add public access setting for content
When public access is turned on LTI authentication is not enforced,
content is accessible to to world, and learners can share URLs and
they'll work without having to launch through the VLE.

LTI authentication is tried first, so that instructors get the dashboard
view instead of the student view. Cookies are searched next to look for
access tokens to an active sessions. Finally public access is used if
the resource has it enabled.