Optionally use Docker to run Chirun; use temp directories for building
If the CHIRUN_DOCKER_IMAGE setting is given, then the process to build a
package runs in Docker using that image. Otherwise, it's assumed that
the `chirun` command is available locally.

This also changes the build process to copy source files to temporary,
directories and to put build output in a temporary directory before
copying it to MEDIA_ROOT.
This avoids filling the editable source directory with incidental files,
and doesn't reveal their local paths in the build log.