set permissions and ownership on the package output directory
When building with docker, the output files end up belonging to root, so we
have to use docker again to change the owner to the owner of the final
output path's parent.

For some reason the directory doesn't get global read+execute
permissions, so this commit adds a line to do that too.