A few things
You can replace or rename files in the file editor.

The deep link shows packages in the same context, then packages in the
same tool, then anything else. (This will have to be refined further in
real use, since "same tool" will be the entire uni, and "other stuff"
would have to be filtered to stuff that is set by an admin to be visible
to everyone)

A few style changes: nav links have no underline; links in the header
nav are contrasty; more spacing around buttons.