Packages have a separate edit UUID
The `edit_uid` field is used for all views that could edit a package. So
knowing `edit_uid` is equivalent to having edit access to a package.

This means that people using the public builder don't need to create an
account, and we don't need to think about how to carry credentials from
an LTI launch - a package created in an LTI context is available for all
deep link launches in that context, and an edit link opening in a new
tab doesn't need to set any session data.

This commit also adds some global template context variables, and
ensures that build progress URLs for packages check the package's
edit_uid in addition to the build's PK.