deep link: handle packages that aren't ready to use
fixes #10

After creating a new package, or selecting a package that hasn't been
successfully built, a holding message is shown directing you to start a
build if there isn't already one in progress.
Once the package has been built, there's a link to continue with the
deep link process.

This is quite clunky - it could start a build for you automatically in
some circumstances, and could maybe automatically move to the next step
when the build is complete. For now, I just want it to work in a way
that's easy to reason about.