remove the deep link configuration page
Working on improving the deep link flow. This exact arrangement probably
won't last long!

But this commit removes the overridden version of the package config
view for inside a deep link launch. Now, you're shown a flash message
that the package has been created and otherwise it's the same as if
you'd clicked on an existing package.