Starting work on cloning packages from git
When creating a package, you can choose to give the URL of a git
repository to clone from.

Cloning and updating is handled in huey tasks.

Getting the list of branches and last commit are handled by running `git
branch` and `git log` in the request handler.

Getting the checked out branch is done by looking at `.git/HEAD` and
expecting it to be of the form `ref: <REF>`.

There's no error handling yet - there needs to be!

working towards #15