Handle launch errors nicely
Until now, launch errors have produced a 500 error response, with no
message displayed.

The platform can send an error message as part of the launch. This can
happen in Canvas when you press the browser back button - it tries to do
the launch again, but sends an error message saying the launch has
already been consumed.

This commit adds some error handling to the LTIView, to display error
messages like these in a nice template.