Support html deep link type
The "html" deep link content type lets us return a fragment of HTML code
that the platform should display.

This commit changes the deep link flow to check the content types that
the platform says it will accept, and offer an "Embedded frame" option.
It returns a div containing a plain link to the chosen item, as well as
an iframe embedding it.