when the build finishes, cope with a missing package name or author, and write any errors to the end of the log
For packages which don't have a title in the config.yml, the build
process was failing when it tried to fill the package's name in the
database with the title from the built manifest.

This changes it to use the current name in the DB if the manifest
doesn't have one, and the same for the author.

The call to save to the database is also wrapped in a try/except block
which will append any error message to the build log, so this kind of
error can be identified.